Fixed added ports having null connections.
Previously, if you tried to override GorkNode and dynamically add Ports to your node, then the connections connected to these nodes are supposed to appear, however a bug caused there to be connections connected to air, which has now been fixed. The issue was basically that the GorkNodeView was being subscribed twice for some reason which caused it to do 2 gork creations in one which resulted in this issue.

There are now also new methods in GorkNode: DeleteInputPort(int index) and DeleteOutputPort(int index)