Functional Dialogue
Created functional dialogue for the DialogueNode and MultiChoiceDialogueNode.
This dialogue system is currently very abstract and requires that you create your own script that derives from GorkDialogueProvider in order for the dialogue to even work.
Example scipts are on the way however so fret not!

This dialogue system is very Coroutine based and will function in almost any scenario you have! Granted you have to make a lot of the code on your own...

Other Notable changes:
- Added a virtual OnStopped method to GorkNode so that the nodes can handle their own logic when being stopped.
- Also added a new GorkScriptingDefineSymbol script which allows the usage of the GORK preprocessor directive. So you can now type "#if GORK" and remove or add code to the compiler based on if you have Gork Graph installed or not. This is really only useful for when you want to create Gork extensions to your own tools.